% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{aaa}
\alias{aaa}
\title{bla 123}
\usage{
aaa()
}
\value{
A keras metric.
}
\description{
Compute AUC score as additional metric. If model has several output neurons with binary crossentropy loss, will use the average score.
}
\examples{
aaa()

}
