% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Add two numbers together}
\usage{
add(x, y, na.rm = FALSE)
}
\arguments{
\item{x}{A number}

\item{y}{Another number}

\item{na.rm}{Drop NAs?}
}
\description{
Add two numbers together
}
\examples{
add(1,2)

}
