% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add2}
\alias{add2}
\title{Returns the sum of two numbers}
\usage{
add2(a, b, c)
}
\arguments{
\item{a}{Number one.}

\item{b}{Number two.}

\item{c}{Number three.}
}
\value{
Sum of two number.
}
\description{
Returns the sum of two numbers
}
\seealso{
\itemize{
\item \code{\link[=enframe]{enframe()}} converts a named vector into a two-column tibble (names and
values).
\item \link{name-repair} documents the details of name repair.
}
}
