% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{methods}
\name{annotation,testS4-method}
\alias{annotation,testS4-method}
\title{Annotation accessor for testS4}
\usage{
\S4method{annotation}{testS4}(object, ...)
}
\arguments{
\item{object}{blah}

\item{...}{Additional arguments ignored}
}
\description{
Annotation accessor for testS4
}

