% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_funcs.R
\name{calc_fibonacci}
\alias{calc_fibonacci}
\title{returns the nth Fibonacci number}
\usage{
calc_fibonacci(n)
}
\arguments{
\item{n}{numeric vector of length 1}
}
\value{
numeric vector of length 1: nth Fibonacci number
}
\description{
\code{calc_fibonacci} returns the nth Fibonacci number
}
