% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{calculate_factorial}
\alias{calculate_factorial}
\title{Calculate Factorial}
\usage{
calculate_factorial(n)
}
\arguments{
\item{n}{An integer for which the factorial will be calculated.}
}
\value{
The factorial of the input integer.
}
\description{
This function calculates the factorial of a given integer.
}
\details{
This function uses a recursive approach to calculate the factorial.
}
\examples{
\dontrun{
if(interactive()){
  calculate_factorial(5)
  }
}
}
