\name{char2helloworld}
\alias{char2helloworld}
\title{Create an instance of the helloworld class.}
\usage{
  char2helloworld(x)
}
\arguments{
  \item{x}{Any character object that you want to convert to
  the helloworld class.}
}
\value{
  Returns \code{x} as an instance of the \code{helloworld}
  class.
}
\description{
  Create an instance of the helloworld class.
}

