% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{count_words}
\alias{count_words}
\title{FUNCTION_TITLE}
\usage{
count_words(text)
}
\arguments{
\item{text}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
