% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubicar.R
\name{cubicar}
\alias{cubicar}
\title{Llevar un numero a la tercera potencia}
\usage{
cubicar(n)
}
\arguments{
\item{n}{numeric(1) El numero}
}
\value{
Un numero cubicado
}
\description{
Llevar un numero a la tercera potencia
}
\examples{

cubicar(2)

}
