% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{custom_model}
\alias{custom_model}
\title{Custom model}
\usage{
custom_model()
}
\value{
A keras model.
}
\description{
Custom model Custom model Custom model Custom model.
}
\examples{
\dontshow{if (reticulate::py_module_available("tensorflow")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

2+2
\donttest{
library(keras)
model <- custom_model()
}
\dontshow{\}) # examplesIf}
}
