% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{dat}
\alias{dat}
\title{dat}
\format{A vector containing 200 samples
\describe{
  \item{x}{x}
  \item{y}{y=x+1}
  ...
}}
\source{
\url{null}
}
\usage{
dat
}
\description{
data set contains 200 randomly generated samples from mixture normal
}
\keyword{datasets}
