% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduct.R
\name{deduct}
\alias{deduct}
\title{Deduct one number from another}
\usage{
deduct(x, y, na.rm = FALSE)
}
\arguments{
\item{x}{A number}

\item{y}{Another number}

\item{na.rm}{Drop NAs?}
}
\description{
Deduct one number from another
}
