% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\docType{data}
\name{demo_table}
\alias{demo_table}
\title{Dummy Table}
\format{
An object of class \code{data.frame} with 3 rows and 2 columns.
}
\usage{
demo_table
}
\description{
\code{demo_table} provides a tiny dummy data set
}
\details{
This data set was included for demo purposes.
}
\examples{
demo_table

}
\keyword{datasets}
