% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{dgp}
\alias{dgp}
\title{Generate draws from simple linear regression}
\usage{
dgp(sample_size, intercept, slope)
}
\arguments{
\item{sample_size}{The sample size}

\item{intercept}{The regression intercept}

\item{slope}{The regression slope}
}
\value{
Returns a dataframe with columns x and y where y is
generated as a linear function of x plus normal random noise, and
x is uniform.
}
\description{
Generate draws from simple linear regression
}
\examples{
dgp(sample_size = 10, intercept = 1, slope = 2)
}
