% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one.R
\name{displaylm}
\alias{displaylm}
\title{Fit and display a linear fit.}
\usage{
displaylm(formula, df)
}
\arguments{
\item{formula}{The \code{lm} standard formula interface.}

\item{df}{The df of data to fit.}
}
\value{
The output of \code{arm::display}.
}
\description{
\code{displaylm} fits a formula to data and then displays it.
}
\details{
\code{displaylm} uses \code{lm} and \code{display} to display the
results of a fit.
}
\examples{
library(testpkg)
library(arm)
alpha = 1
beta <- 0.7
df <- data.frame(
  x = 1:10,
  y = alpha + beta * 1:10 + rnorm(10))
displaylm(formula = 'y ~ 1 + x', df = df)
}
