% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_funcs.R
\name{extract_string_element}
\alias{extract_string_element}
\title{extracts the nth element from a string}
\usage{
extract_string_element(string1, n)
}
\arguments{
\item{string1}{character vector of length 1}

\item{n}{numeric vector of length 1}
}
\value{
character vector of length 1 containing a single character
}
\description{
\code{extract_string_element} extracts the nth element from a string
}
