% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{fahrenheit_to_celsius}
\alias{fahrenheit_to_celsius}
\title{FUNCTION_TITLE}
\usage{
fahrenheit_to_celsius(temp_f)
}
\arguments{
\item{temp_f}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
