% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-filter.R
\name{filter}
\alias{filter}
\title{Use the \code{dplyr::filter()} S3 generic}
\usage{
filter(.data, ..., .preserve = FALSE)
}
\description{
See \code{dplyr::\link[dplyr]{filter}} for details.
}
\examples{
filter(iris, Species == "versicolor")
filter(iris, Species == "versicolor" & Petal.Width < 4)
}
\keyword{internal}
