% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{height_handspan}
\alias{height_handspan}
\title{Height and hanspan for Econ 103 Students}
\format{A data frame with 69 rows and 2 variables:
\describe{
  \item{height}{height, in inches}
  \item{handspan}{handspan, in centimeters}
}}
\source{
\url{http://ditraglia.com/econ103/survey_clean.csv/}
}
\usage{
height_handspan
}
\description{
A dataset containing the height and handspan of Econ 103 Students
}
\keyword{datasets}
