% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Greets a person}
\usage{
hello(name)
}
\arguments{
\item{name}{Character. A name of a person.}
}
\value{
A message saying hello
}
\description{
Greets a person
}
\examples{
\dontrun{
hello("Hisham M")
hello("Jonas")
}
}
