% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello-s3.R
\name{hello_s3}
\alias{hello_s3}
\alias{hello_s3.character}
\alias{hello_s3.numeric}
\alias{hello_s3.virus_df}
\title{Greets a person or a number}
\usage{
hello_s3(x, ...)

\method{hello_s3}{character}(x, ...)

\method{hello_s3}{numeric}(x, ...)

\method{hello_s3}{virus_df}(x, ...)
}
\arguments{
\item{name}{Character or numeric. A name of a person or a number from 1 to 3}
}
\value{
A message saying hello
}
\description{
Greets a person or a number
}
\examples{
hello_s3("Hisham")
hello_s3("Jonas")
hello_s3(1)
hello_s3(2)
df <- data.frame(virus_name = c("herpes", "influenza", "covid"))
virus_df <- new_virus_df(df)
hello_s3(virus_df)
}
