% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{indicate_concerns}
\alias{indicate_concerns}
\title{Silly Warning}
\usage{
indicate_concerns()
}
\description{
\code{indicate_concerns} prompts a silly warning to the console
}
\details{
This function was included for demo purposes.
It is used as a dummy function with some elaborate
documentation.
}
\examples{
indicate_concerns()

}
