% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_release.R
\name{init_release}
\alias{init_release}
\alias{complete_release}
\title{Release a New \code{somaverse} Package Version}
\usage{
init_release(ver = NULL, check = FALSE)

complete_release(tag = NULL)
}
\arguments{
\item{ver}{Character. See \code{usethis::use_version()}.}

\item{check}{Logical. Should \code{devtools::check()} be run internally?}

\item{tag}{Character. The name of the version tag for the release, e.g. "v3.2.1"}
}
\value{
The values of \code{usethis::sitrep()} and \code{git2r::status()} are returned.
}
\description{
Go through a release checklist to perform steps necessary for a \code{somaverse}
package release.
}
\section{Functions}{
\itemize{
\item \code{complete_release}: Complete the package release.
}}

\examples{
\dontrun{
soma_release()
}
}
