\name{initialize}
\alias{initialize}
\title{Create an instance of the helloworld class using new/initialize.}
\description{
  This is a dummy method for demonstration purposes, to
  produce an instance of the dummy `helloworld' class. Used
  internally in the `testpkg' package.
}
\seealso{
  initialize
}
\keyword{internal}

