% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{is_even}
\alias{is_even}
\title{FUNCTION_TITLE}
\usage{
is_even(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
