\name{levels.no}
\alias{levels.no}
\alias{f}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
test levels.no
}
\description{
A description
}
\usage{
levels.no(xx, ...)
f(x)
}
\arguments{
  \item{x, xx}{
arg 1
}
  \item{\dots}{
others
}
}
\examples{
x <- structure(1, class = "no")
levels(x)
f
f(x)
}
