% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meansd.R
\name{meansd}
\alias{meansd}
\title{Hacer mean y sd sobre un vector}
\usage{
meansd(variable, na.rm = TRUE, nombre = NULL, digitos = 2)
}
\arguments{
\item{variable}{Un vector numerico}

\item{na.rm}{(TRUE) eliminar valores NA cuando resumiendo los datos}

\item{nombre}{(opcional) sequencia de caracteres para el nombre de el
vector \code{variable}}

\item{digitos}{Mostrar este numero de digitos en la salida}
}
\description{
Toma un variable numerico y devuelve el promedio \code{mean} y
desviacion estandar \code{sd}
}
\examples{

data(mtcars)

meansd(mtcars$mpg)

}
