% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{multiply}
\alias{multiply}
\title{FUNCTION_TITLE}
\usage{
multiply(a, b)
}
\arguments{
\item{a}{PARAM_DESCRIPTION}

\item{b}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
