\name{my.mean}
\alias{my.mean}
\title{Calculate the arithmetic mean}
\description{
Calculate the arithmetic mean of numeric vector
}
\usage{
my.mean(x)
}
\arguments{
  \item{x}{numeric verctor}
}
\details{
Calculate the arithmetic mean of numeric vector
}
\value{numeric}
\author{Ryu Choonghyun}
\seealso{\code{\link{center}}}
\examples{
my.mean(1:10)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
