% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_funcs.R
\name{my_add}
\alias{my_add}
\title{adds two numbers}
\usage{
my_add(x, y)
}
\arguments{
\item{x}{numeric vector of length 1}

\item{y}{numeric vector of length 1}
}
\value{
numeric vector of length 1
}
\description{
\code{my_add} adds two numbers
}
