% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_function.R
\name{my_function}
\alias{my_function}
\title{My First Function}
\usage{
my_function(x, y)
}
\arguments{
\item{x}{A numeric value.}

\item{y}{A numeric value.}
}
\value{
The sum of x and y.
}
\description{
This function adds two numbers.
}
