% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Binomial Plot Simulation}
\usage{
mybin(iter, n, p, ret)
}
\arguments{
\item{iter, }{the number of iterations in the function}

\item{n}{, the sample size}

\item{p}{, the probability of the sample}

\item{ret}{, boolean value in regards to data output}
}
\value{
, a simulated binomial experiment
}
\description{
Takes the number of iterations (iter), the sample size (n), probability (p), and a boolean (ret).
The boolean, ret, is used to signify whether you would prefer you would like extra information output
}
\examples{
mybin(iter=5, n=10, p=.7, ret=TRUE)

}
