% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{Histogram of the sum of columns}
\usage{
myclt(n, iter)
}
\arguments{
\item{n}{, number of samples}

\item{iter}{, number of columns}
}
\value{
, Histograph of the sum of columns of the calculated sample
}
\description{
Takes in the number of columns and number of samples. Using these parameters it creates a matrix of random samples. The columns of
this matrix is summed and then displayed in a histogram to display the central limit theorm.
}
\examples{
myclt(10, 5)

}
