% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycyi.R
\name{mycyi}
\alias{mycyi}
\title{Confidence Intervals for sample distributions}
\usage{
mycyi(x)
}
\arguments{
\item{x}{Sample distribution}
}
\value{
95 percent confidence interval  for the sample in the parameters
}
\description{
This function takes in a sample distribution and return a 95 percent confidence interval
}
\examples{
set.seed(23);
x = rnorm(30,mean=10,sd=12);
mycyi(x)


}
