% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun.R
\name{myfun}
\alias{myfun}
\title{myfun}
\usage{
myfun(a = 1, b = 2, c)
}
\arguments{
\item{a}{First argument}

\item{b}{Second argument}

\item{c}{Third Argument}
}
\value{
a + b + c
}
\description{
myfun
}
\examples{
myfun(1, 2, 3)
}
