% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunction.R
\name{myncurve}
\alias{myncurve}
\title{Normal curve with highlighted area}
\usage{
myncurve(mu, sigma, xt)
}
\arguments{
\item{mu}{, mean}

\item{sigma}{, standard deviation}

\item{xt}{, x axis of the graph}
}
\value{
, normally distributed curve with the region between the curve and xt being shaded along with the area of the region.
}
\description{
Uses the parameters to generate a normal curve, and then uses the parameter xt to shade the region below the curve from 0 to xt.
}
\examples{
myncurve(mu=10, sigma=5, xt =5)


}
