% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplot.R
\name{myplot}
\alias{myplot}
\title{A placeholder function using roxygen}
\usage{
myplot(...)
}
\arguments{
\item{...}{Passed to plot.}
}
\value{
A plot
}
\description{
A thin wrapper around base plot.
}
\examples{
myplot()
}
