% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_funcs.R
\name{plot_line}
\alias{plot_line}
\title{makes a line plot}
\usage{
plot_line(data_frame, x_name, y_name)
}
\arguments{
\item{data_frame}{data frame with at least columns x_name and y_name}

\item{x_name}{character vector of length 1: name of variable to plot on x axis}

\item{y_name}{character vector of length 1: name of variable to plot on y axis}
}
\value{
ggplot object
}
\description{
\code{plot_line} takes a data frame and makes a line plot with two of the 
columns, one on each axis.
}
