% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{print_three}
\alias{print_three}
\title{Print three
an example function that always returns 3}
\usage{
print_three()
}
\value{
3
}
\description{
Print three
an example function that always returns 3
}
\examples{
print_three()
}
