% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{product}
\alias{product}
\title{Product function}
\usage{
product(a, b)
}
\arguments{
\item{a}{number}

\item{b}{number}
}
\value{
product of two numbers
}
\description{
Finds the product of two numbers
}
\examples{
product(1,2)
product(4,9)
}
