% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.r
\name{euclidean}
\alias{euclidean}
\title{euclidean function}
\usage{
euclidean(a,b)
}
\arguments{
\item{a}{A number.}

\item{b}{A number.}
}
\value{
the greatest divisor of \code{a} and \code{b}
}
\description{
euclidean function
}
\details{
nothing
}
\section{Functions}{
\itemize{
\item \code{euclidean}: the greatest divisor of a and b. see\url{http://https://en.wikipedia.org/wiki/Euclidean_algorithm} for more details.
}}

\examples{
euclidean(10,20)
euclidean(112,4522)
}
