% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_funcs.R
\name{reverse_string}
\alias{reverse_string}
\title{reverses a string}
\usage{
reverse_string(my_string)
}
\arguments{
\item{my_string}{character vector}
}
\value{
character vector
}
\description{
\code{reverse_string} reverses a string
}
