% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_gz.r
\name{search_gz}
\alias{search_gz}
\title{Scan a gzipped file for text}
\usage{
search_gz(terms, path)
}
\arguments{
\item{terms}{Character vector; search terms (most likely GenBank accession
numbers)}

\item{path}{Path to the gzipped file to scan}
}
\value{
Logical
}
\description{
Scans a zipped file for text
strings and returns TRUE if any are present.
}
