% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepMillier.R
\name{sepMillier}
\alias{sepMillier}
\title{Sep Millier}
\usage{
sepMillier(x)
}
\arguments{
\item{x}{numeric value}
}
\value{
string value of the input with space thousand separator
}
\description{
Sep Millier
}
\examples{
sepMillier(28000)
# output: 28 000
}
