% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{signal}
\alias{signal}
\alias{value}
\alias{signal_done}
\alias{signal_todo}
\alias{signal_oops}
\alias{signal_info}
\title{Report a Value to GUI}
\usage{
value(x)

signal_done(...)

signal_todo(...)

signal_oops(...)

signal_info(...)
}
\arguments{
\item{x, ...}{Character. Each element of the vector becomes an entry
in a comma separated list and a \code{blue} color is added. Or \code{...} passed
directly to \code{\link[=cat]{cat()}}.}
}
\description{
Similar to \verb{usethis::ui_*()} but does not require importing the
\pkg{usethis} package.
}
\section{Functions}{
\itemize{
\item \code{value}: Signal a value to the UI. Similar to
\code{\link[usethis:ui]{usethis::ui_value()}}.

\item \code{signal_done}: Signal a completed task to the UI. Similar to
\code{\link[usethis:ui]{usethis::ui_done()}}.

\item \code{signal_todo}: Signal a to-do task to the UI. Similar to
\code{\link[usethis:ui]{usethis::ui_todo()}}.

\item \code{signal_oops}: Signal oops error to the UI. Similar to
\code{\link[usethis:ui]{usethis::ui_oops()}}.

\item \code{signal_info}: Signal info to the UI. Similar to
\code{\link[usethis:ui]{usethis::ui_info()}}.
}}

