\docType{methods}
\name{sillyadd}
\alias{sillyadd}
\alias{sillyadd,ANY-method}
\alias{sillyadd,character-method}
\title{Trivial addition of two arguments with S4 method}
\usage{
  sillyadd(x, y)
}
\arguments{
  \item{x}{Description of \code{x}.}

  \item{y}{Description of \code{y}.}
}
\value{
  The two arguments added together.
}
\description{
  Some additional details go here.
}
\examples{
##
}

