\name{sillysubtract}
\alias{sillysubtract}
\title{Trivial subtraction of y from x}
\usage{
  sillysubtract(x, y)
}
\arguments{
  \item{x}{Description of \code{x}.}

  \item{y}{Description of \code{y}.}
}
\value{
  The result of \code{x - y}.
}
\description{
  Some additional details go here.
}
\examples{
##
}

