% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{square}
\alias{square}
\title{square a number}
\usage{
square(x)
}
\arguments{
\item{x}{a number to be squared}
}
\value{
a number
}
\description{
computes the square of a number, i.e. the number multiplied by itself
}
\examples{
square(10)
square(4)
}
