% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{square_say}
\alias{square_say}
\title{Title}
\usage{
square_say(x)
}
\arguments{
\item{x}{a number to be squared by cow}
}
\value{
a long output with a cow saying the answer
}
\description{
Title
}
\examples{
square_say(10)

}
