% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{square_val}
\alias{square_val}
\title{Square Value}
\usage{
square_val(x)
}
\arguments{
\item{x}{numeric value}
}
\value{
squared value
}
\description{
Square Value
}
\examples{
square_val(2)
# output: 4
square_val(9)
# output: 81
}
