% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{summary_stats}
\alias{summary_stats}
\title{FUNCTION_TITLE}
\usage{
summary_stats(data, column)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{column}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
