% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_funcs.R
\name{teenage_capitalise}
\alias{teenage_capitalise}
\title{randomly capitalises letters in a string}
\usage{
teenage_capitalise(my_string)
}
\arguments{
\item{my_string}{character vector}
}
\value{
character vector
}
\description{
\code{teenage_capitalise} randomly capitalises letters in a string
}
