% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test.R, R/data.R
\docType{package}
\name{Test-package}
\alias{Test}
\alias{Test-package}
\title{Test: A Testing Ground}
\description{
Just a test package to debug devtools and builds.

Just a test package to debug devtools and builds.
}
\seealso{
Useful links:
\itemize{
  \item \url{www.google.com}
}


Useful links:
\itemize{
  \item \url{www.google.com}
}

}
\author{
\strong{Maintainer}: Stu Field \email{sfield@somalogic.com}


\strong{Maintainer}: Stu Field \email{sfield@somalogic.com}

}
\keyword{internal}
\keyword{package}
\keyword{test}
