% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_fun.R
\name{test_fun}
\alias{test_fun}
\title{Test}
\usage{
test_fun(x)
}
\arguments{
\item{x}{a value}
}
\value{
x pasted together
}
\description{
Just a test function
}
\examples{
test_fun(2)
test_fun(seq(1, 4))
test_fun("Yo!")
}
