\name{test_pkg}
\alias{test_pkg}

\title{
A simple test function
}
\description{
A placeholder function for testing pkgdown support of Quarto articles.
The function does not do anything useful. 
}
\usage{
test_pkg <- function()
}
\arguments{

}
\details{
The pkgdown package apparently supports Quarto articles.
This function is part of a shell package meant to test the support.
The function takes no arguments and only returns a message to the user. 
}

\value{
None
}



\examples{
test_pkg()
}

