% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testfunc.R
\name{testfunc}
\alias{testfunc}
\title{Test function that prints a message.}
\usage{
testfunc()
}
\value{
Single logical \code{TRUE}.
}
\description{
This is just to test out the suppression of messages.
}

