% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{testing}
\alias{testing}
\title{Another test function}
\usage{
testing(a, b)
}
\arguments{
\item{a, b}{character(1) A letter, either lowercase or uppercase}
}
\value{
Another message
}
\description{
Another test function
}
