% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubicar.R
\name{testLines}
\alias{testLines}
\title{Contar el numero de linas en una carpeta}
\usage{
testLines(testFolder)
}
\arguments{
\item{testFolder}{La carpeta de pruebas de unidad}
}
\value{
Un numero correspondiendo a el conteo de lineas dentro la carpeta
}
\description{
Contar el numero de linas en una carpeta
}
\examples{

## NOTA (OJO) En paquetes no usamos rutas fuera del paquete.
## Esto es solo un ejemplo

testFolder <- "~/GenomicRanges/inst/unitTests"
testLines(testFolder)

}
