% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestPkg-package.R, R/data.R
\docType{package}
\name{TestPkg-package}
\alias{TestPkg}
\alias{TestPkg-package}
\title{TestPkg: A Testing Ground}
\description{
Just a test package to debug devtools and builds.

Just a test package to debug devtools and builds.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/stufield}
}


Useful links:
\itemize{
  \item \url{http://github.com/stufield}
}

}
\author{
\strong{Maintainer}: Stu Field \email{stu.g.field@gmail.com}


\strong{Maintainer}: Stu Field \email{stu.g.field@gmail.com}

}
\keyword{internal}
\keyword{package}
\keyword{test}
