% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top.R
\name{top}
\alias{top}
\title{Show the top part of an object}
\usage{
top(mat, n = 5)
}
\arguments{
\item{mat}{a matrix or data.frame}

\item{n}{Number of rows/columns to display}
}
\description{
This combines head with a number of codes
}
\examples{
\dontrun{
# I don't want you to run this
}
x = matrix(rnorm(100), ncol = 10)
Sys.sleep(6)
top(x)
}
\seealso{
\code{\link[utiles]{head}}
}
