% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valueUnite.R
\name{valueUnite}
\alias{valueUnite}
\title{Value Unite}
\usage{
valueUnite(val)
}
\arguments{
\item{val}{numeric value}
}
\value{
string value of the input with ISU abbreviation
}
\description{
Value Unite
}
\examples{
valueUnite(2800)
# output: 2.8 K
valueUnite(2800000)
# output: 2.8 M
valueUnite(2800000000)
# output: 2.8 MDS
}
