% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/your_name.R
\name{your_name}
\alias{your_name}
\title{Test Function}
\usage{
your_name(x)
}
\arguments{
\item{x}{A string}
}
\description{
This function is for testing package building.
}
\examples{
your_name("Derek")
}
\keyword{misc}
