% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{zz}
\alias{zz}
\title{bla 333}
\usage{
zz()
}
\value{
A keras layer.
}
\description{
zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz.
}
\examples{
\dontshow{if (reticulate::py_module_available("tensorflow")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

\donttest{
library(keras)
l <- zz()
}
\dontshow{\}) # examplesIf}
}
