% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_api.R
\name{github_api}
\alias{github_api}
\title{Hit the GitHub API}
\usage{
github_api(path)
}
\arguments{
\item{path}{String of the endpoint destination}
}
\description{
Hit the GitHub API
}
\examples{
# My GitHub user account
tryCatch(github_api("/users/stufield"), error = function(e) e)

# This package repo
tryCatch(github_api("/repos/stufield/TestPkg"), error = function(e) e)
}
\author{
Stu Field
}
