\name{HWplot}
\alias{HWplot}
\title{Create a helloworld-ified plot using qplot arguments.}
\usage{
  HWplot(...)
}
\arguments{
  \item{...}{Comma separated list of named arguments to
  qplot.}
}
\value{
  A helloworld-ified qplot. Oh, you'll see.
}
\description{
  Create a helloworld-ified plot using qplot arguments.
}
\seealso{
  qplot
}

