% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFun.R
\name{myFun}
\alias{myFun}
\title{Squares a number}
\usage{
myFun(x = 1)
}
\arguments{
\item{x}{\code{numeric(1)} A numeric vector of numbers to square.}
}
\value{
\code{numeric(1)}
}
\description{
This function accepts and numeric vector and returns it's value squared.
}
