% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script1.R
\name{pkg_inst}
\alias{pkg_inst}
\title{Calls pkgdepends::new_pkg_installation_proposal}
\usage{
pkg_inst(refs, snapshot, destdir, config = list())
}
\arguments{
\item{refs}{Package references.}

\item{snapshot}{MRAN snapshot date.}

\item{destdir}{Directory where packages are to be installed.}

\item{config}{Additional configuration options.}
}
\description{
Calls pkgdepends::new_pkg_installation_proposal
}
