\name{Testpkg-package}
\alias{Testpkg-package}
\alias{Testpkg}
\docType{package}
\title{
User defined packages
}
\description{
Creation Packages Example
}
\details{
\tabular{ll}{
Package: \tab Testpkg\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-08-11\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Packages example by Ryu Choonghyun

Maintainer: Ryu Choonghyun <bdboys@naver.com>
}
\keyword{ package }
