\name{testPkg-package}
\alias{testPkg-package}
\alias{testPkg}
\docType{package}
\title{\packageTitle{testPkg}}
\description{\packageDescription{testPkg}}
\section{Package Content}{\packageIndices{testPkg}}
\author{\packageAuthor{testPkg}}
\section{Maintainer}{\packageMaintainer{testPkg}}
\keyword{package}
