\name{testpkg-package}
\alias{testpkg-package}
\alias{testpkg}
\docType{package}
\title{\packageTitle{testpkg}}
\description{\packageDescription{testpkg}}
\section{Package Content}{\packageIndices{testpkg}}
\author{\packageAuthor{testpkg}}
\section{Maintainer}{\packageMaintainer{testpkg}}
\keyword{package}
