% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpkg-package.R
\docType{package}
\name{testpkg-package}
\alias{testpkg}
\alias{testpkg-package}
\title{testpkg: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/klevdiamanti/testpkg}
  \item Report bugs at \url{https://github.com/klevdiamanti/testpkg/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
