% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{testpkg-package}
\alias{testpkg}
\alias{testpkg-package}
\title{pkg bla}
\description{
pkg is an open source software library for building deep neural
networks for genomic modeling.
}
\author{
\strong{Maintainer}: René Mreches \email{rene.mreches@helmholtz-hzi.de}

}
